+++ config/index.ts
@@ -1,3 +1,5 @@
 export const defaultTablePath = './table.xlsx'
 export const convertedExcelName = 'converted.xlsx'
 export const convertedTablePath = `./${convertedExcelName}`
+export const statisticsExcelName = 'statistics.xlsx'
+export const statisticsExcelPath = `./${statisticsExcelName}`