+++ src/commands/convert.ts
@@ -1,9 +1,8 @@
-import * as path from 'path'
 import * as inquirer from 'inquirer'
 import * as xlsx from 'xlsx'
 import { defaultTablePath, convertedExcelName } from '../../config'
 
-import { locateFile, locateDir } from '../helpers'
+import { locateFile, exportWorkbook } from '../helpers'
 import convertSheet from './convertSheet'
 import Mfcdm from '..'
 
@@ -22,13 +21,6 @@ const askWhichSheets = async (sheetNames: string[]) => {
 
   return inquirer.prompt(questions)
     .then((answers: inquirer.Answers) => answers.sheets)
-}
-
-const exportWorkbook = async (workbook: xlsx.WorkBook) => {
-  const directory = await locateDir(`Where should the ${convertedExcelName} file be put?`)
-  const absolute = path.resolve(directory)
-  const fullPath = path.join(absolute, convertedExcelName)
-  xlsx.writeFile(workbook, fullPath)
 }
 
 /* ============================== */
@@ -63,13 +55,13 @@ const convert = async (mfcdm: Mfcdm) => {
 
   for (let sheetName of sheetsToConvert) {
     const sheet = workbook.Sheets[sheetName]
-    const sheetData = xlsx.utils.sheet_to_json(sheet)
+    const sheetData: {}[] = xlsx.utils.sheet_to_json(sheet)
     const convertedSheetData = await convertSheet(sheetName, sheetData, middleware[sheetName])
     const worksheet = xlsx.utils.json_to_sheet(convertedSheetData)
     xlsx.utils.book_append_sheet(outputWorkbook, worksheet, sheetName)
   }
 
-  exportWorkbook(outputWorkbook)
+  await exportWorkbook(outputWorkbook, convertedExcelName)
 }
 
 export default convert