+++ src/helpers/index.ts
@@ -1,2 +1,3 @@
 export { default as locateDir } from './locateDir'
 export { default as locateFile } from './locateFile'
+export { default as exportWorkbook } from './exportWorkbook'